requirements.txt
absl-py==0.11.0
appdirs==1.4.4
appnope==0.1.2
astunparse==1.6.3
backcall==0.2.0
black==20.8b1
cachetools==4.2.0
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
decorator==4.4.2
farasapy==0.0.11
Flask==1.1.2
flatbuffers==1.12
gast==0.3.3
google-auth==1.24.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
grpcio==1.32.0
gunicorn==19.9.0
h5py==2.10.0
idna==2.10
ipdb==0.13.4
ipython==7.19.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.18.0
Jinja2==2.11.2
joblib==1.0.0
Keras-Preprocessing==1.1.2
langdetect==1.0.8
Markdown==3.3.3
MarkupSafe==1.1.1
mypy-extensions==0.4.3
nltk==3.5
numpy==1.19.4
oauthlib==3.1.0
opt-einsum==3.3.0
opencv==3.1
parso==0.8.1
pathspec==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.8
protobuf==3.14.0
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.7.3
regex==2020.11.13
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.6
scikit-learn==0.24.0
scipy==1.5.4
sentencepiece==0.1.94
six==1.15.0
sklearn==0.0
tensorboard==2.4.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.4.0
tensorflow-estimator==2.4.0
termcolor==1.1.0
threadpoolctl==2.1.0
tkseem==0.0.3
tnkeeh==0.0.3
toml==0.10.2
tqdm==4.55.0
traitlets==5.0.5
typed-ast==1.4.1
typing-extensions==3.7.4.3
urllib3==1.26.2
wcwidth==0.2.5
Werkzeug==1.0.1
wrapt==1.12.1
